+++ src/renderer/entities/governance/lib/governanceService.ts
@@ -8,7 +8,6 @@ import {
 import { type BN, BN_ZERO } from '@polkadot/util';
 
 import {
-  type AccountVote,
   type Address,
   type Deposit,
   type LinearDecreasingCurve,
@@ -19,16 +18,13 @@ import {
   type SteppedDecreasingCurve,
   type TrackId,
   type TrackInfo,
-  type Voting,
   type VotingCurve,
 } from '@/shared/core';
 
 export const governanceService = {
   getReferendums,
   getReferendum,
-  getVotingFor,
   getTrackLocks,
-  subscribeTrackLocks,
   getTracks,
 };
 
@@ -172,10 +168,10 @@ async function getReferendums(api: ApiPromise): Promise<Referendum[]> {
   return result;
 }
 
-// function subscribeReferendumsFx(api: ApiPromise, callback: (referendums?: Referendum[]) => void): Promise<() => void> {
+// async function subscribeReferendumsFx(api: ApiPromise, callback: (referendums?: Referendum[]) => void): Promise<() => void> {
 //   console.log(1, 'referendums');
 
-//   return api.query.referenda.referendumInfoFor(1, (referendums: any) => {
+//   return api.query.referenda.referendumInfoFor.entries((referendums: any) => {
 //     console.log('2, referendums', referendums);
 //     try {
 //       const result: Referendum[] = [];
@@ -209,101 +205,6 @@ async function getReferendum(id: ReferendumId, api: ApiPromise): Promise<Referen
   return mapReferendum(id, palette.unwrap());
 }
 
-async function getVotingFor(
-  api: ApiPromise,
-  tracksIds: TrackId[],
-  addresses: Address[],
-  callback: (res?: Record<Address, Record<TrackId, Voting>>) => void,
-): Promise<() => void> {
-  const tuples = addresses.flatMap((address) => tracksIds.map((trackId) => [address, trackId]));
-
-  return api.query.convictionVoting.votingFor.multi(tuples, (votings) => {
-    const result = addresses.reduce<Record<Address, Record<TrackId, Voting>>>((acc, address) => {
-      acc[address] = {};
-
-      return acc;
-    }, {});
-
-    for (const [index, convictionVoting] of votings.entries()) {
-      if (convictionVoting.isStorageFallback) continue;
-
-      const address = tuples[index]?.[0];
-      const trackId = tuples[index]?.[1];
-      if (!address || !trackId) {
-        continue;
-      }
-
-      if (convictionVoting.isDelegating) {
-        const delegation = convictionVoting.asDelegating;
-
-        result[address][trackId] = {
-          type: 'Delegating',
-          address,
-          track: trackId,
-          balance: delegation.balance.toBn(),
-          conviction: delegation.conviction.type,
-          target: delegation.target.toString(),
-          prior: {
-            unlockAt: delegation.prior[0].toNumber(),
-            amount: delegation.prior[1].toBn(),
-          },
-        };
-      }
-
-      if (convictionVoting.isCasting) {
-        const votes: Record<ReferendumId, AccountVote> = {};
-        for (const [referendumIndex, vote] of convictionVoting.asCasting.votes) {
-          const referendumId = referendumIndex.toString();
-
-          if (vote.isStandard) {
-            const standardVote = vote.asStandard;
-            votes[referendumId] = {
-              type: 'Standard',
-              vote: {
-                aye: standardVote.vote.isAye,
-                conviction: standardVote.vote.conviction.type,
-              },
-              balance: standardVote.balance.toBn(),
-            };
-          }
-
-          if (vote.isSplit) {
-            const splitVote = vote.asSplit;
-            votes[referendumId] = {
-              type: 'Split',
-              aye: splitVote.aye.toBn(),
-              nay: splitVote.nay.toBn(),
-            };
-          }
-
-          if (vote.isSplitAbstain) {
-            const splitAbstainVote = vote.asSplitAbstain;
-            votes[referendumId] = {
-              type: 'SplitAbstain',
-              aye: splitAbstainVote.aye.toBn(),
-              nay: splitAbstainVote.nay.toBn(),
-              abstain: splitAbstainVote.abstain.toBn(),
-            };
-          }
-        }
-
-        result[address][trackId] = {
-          type: 'Casting',
-          track: trackId,
-          address,
-          votes,
-          prior: {
-            unlockAt: convictionVoting.asCasting.prior[0].toNumber(),
-            amount: convictionVoting.asCasting.prior[1].toBn(),
-          },
-        };
-      }
-    }
-
-    callback(result);
-  });
-}
-
 async function getTrackLocks(api: ApiPromise, addresses: Address[]): Promise<Record<Address, Record<TrackId, BN>>> {
   const tuples = await api.query.convictionVoting.classLocksFor.multi(addresses);
   const result: Record<Address, Record<TrackId, BN>> = {};
@@ -319,27 +220,6 @@ async function getTrackLocks(api: ApiPromise, addresses: Address[]): Promise<Rec
   }
 
   return result;
-}
-
-function subscribeTrackLocks(
-  api: ApiPromise,
-  addresses: Address[],
-  callback: (res?: Record<Address, Record<TrackId, BN>>) => void,
-): Promise<() => void> {
-  return api.query.convictionVoting.classLocksFor.multi(addresses, (tuples) => {
-    const result: Record<Address, Record<TrackId, BN>> = {};
-
-    for (const [index, locks] of tuples.entries()) {
-      const lockData = locks.reduce<Record<TrackId, BN>>((acc, lock) => {
-        acc[lock[0].toString()] = lock[1].toBn();
-
-        return acc;
-      }, {});
-
-      result[addresses[index]] = lockData;
-    }
-    callback(result);
-  });
 }
 
 function getTracks(api: ApiPromise): Record<TrackId, TrackInfo> {
